gateway 整合 security 异常 未解决
