+++ learn-cloud-gateway/src/main/java/com/billow/gateway/GatewayApp.java
@@ -1,0 +1,19 @@
+package com.billow.gateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.SecurityProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.ComponentScan;
+
+@EnableDiscoveryClient
+@SpringBootApplication
+@ComponentScan("com.billow")
+@EnableConfigurationProperties(SecurityProperties.class)
+public class GatewayApp {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GatewayApp.class, args);
+    }
+}
