+++ learn-cloud-gateway/src/main/java/com/billow/gateway/properties/CustomProperties.java
@@ -1,0 +1,19 @@
+package com.billow.gateway.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+/**
+ * 约定的配置信息
+ *
+ * @author liuyongtao
+ * @create 2019-04-30 9:08
+ */
+@Data
+@Component
+@ConfigurationProperties(prefix = "custom")
+public class CustomProperties {
+
+    private SecurityProperties security = new SecurityProperties();
+}
