+++ learn-cloud-gateway/src/main/java/com/billow/gateway/properties/SecurityProperties.java
@@ -1,0 +1,17 @@
+package com.billow.gateway.properties;
+
+import lombok.Data;
+
+/**
+ * 关于安全的配置
+ *
+ * @author liuyongtao
+ * @create 2019-05-22 14:25
+ */
+@Data
+public class SecurityProperties {
+
+    // 认证服务器的地址
+    private String authServer;
+
+}
