+++ learn-cloud-gateway/src/main/java/com/billow/gateway/redis/RedisClusterProperties.java
@@ -1,0 +1,37 @@
+package com.billow.zuul.redis;
+
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+/**
+ * 集群环境配置
+ *
+ * @author LiuYongTao
+ * @date 2019/7/29 9:12
+ */
+@Component
+@ConfigurationProperties(prefix = "spring.redis.cluster")
+public class RedisClusterProperties {
+
+
+    private String nodes;
+
+    private Integer commandTimeout;
+
+    public String getNodes() {
+        return nodes;
+    }
+
+    public void setNodes(String nodes) {
+        this.nodes = nodes;
+    }
+
+    public Integer getCommandTimeout() {
+        return commandTimeout;
+    }
+
+    public void setCommandTimeout(Integer commandTimeout) {
+        this.commandTimeout = commandTimeout;
+    }
+}
