--- learn-cloud-zuul/src/main/java/com/billow/zuul/ZuulApp.java
@@ -1,22 +1,0 @@
-package com.billow.zuul;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-import org.springframework.context.annotation.ComponentScan;
-
-
-@EnableZuulProxy
-@EnableDiscoveryClient
-@SpringBootApplication
-@ComponentScan("com.billow")
-@EnableConfigurationProperties(SecurityProperties.class)
-public class ZuulApp {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ZuulApp.class, args);
-    }
-}
