--- learn-cloud-zuul/src/main/java/com/billow/zuul/config/Swagger2Config.java
@@ -1,111 +1,0 @@
-package com.billow.zuul.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.OAuthBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.AuthorizationScope;
-import springfox.documentation.service.Contact;
-import springfox.documentation.service.GrantType;
-import springfox.documentation.service.ResourceOwnerPasswordCredentialsGrant;
-import springfox.documentation.service.SecurityReference;
-import springfox.documentation.service.SecurityScheme;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spi.service.contexts.SecurityContext;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger.web.SecurityConfiguration;
-import springfox.documentation.swagger.web.SecurityConfigurationBuilder;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-/**
- * 用于接口文档
- *
- * @author liuyongtao
- * @create 2018-02-08 11:40
- */
-@Configuration
-@EnableSwagger2 // 启用Swagger2
-public class Swagger2Config {
-
-    @Value("${swagger.basepackage}")
-    private String basePackage;
-    @Value("${swagger.service.name}")
-    private String serviceName;
-    @Value("${swagger.service.description}")
-    private String description;
-    @Value("${swagger.service.version}")
-    private String version;
-    @Value("${swagger.service.contact.developer}")
-    private String developer;
-    @Value("${swagger.service.contact.url}")
-    private String url;
-    @Value("${swagger.service.contact.email}")
-    private String email;
-    @Value("${security.oauth2.client.access-token-uri}")
-    private String tokenUrl;
-
-    private static final String CLIENT_ID = "swagger";
-    private static final String CLIENT_SECRET = "swagger";
-    private static final String SCOPE = "webapp";
-
-
-    @Bean
-    public Docket createRestApi() {
-
-        return (new Docket(DocumentationType.SWAGGER_2))
-                .apiInfo(this.apiInfo())
-                .select()
-                .apis(RequestHandlerSelectors.basePackage(this.basePackage))
-                .paths(PathSelectors.regex("^(?!auth).*$"))
-                .build()
-                .securityContexts(Collections.singletonList(securityContext()))
-                .securitySchemes(Collections.singletonList(securityScheme()));
-    }
-
-    @Bean
-    public SecurityConfiguration security() {
-        return SecurityConfigurationBuilder.builder()
-                .clientId(CLIENT_ID)
-                .clientSecret(CLIENT_SECRET)
-                .scopeSeparator(SCOPE)
-                .build();
-    }
-
-    private ApiInfo apiInfo() {
-        return (new ApiInfoBuilder())
-                .title(this.serviceName + " Restful APIs")
-                .description(this.description)
-                .contact(new Contact(this.developer, this.url, this.email))
-                .version(version)
-                .build();
-    }
-
-    private SecurityContext securityContext() {
-        return SecurityContext.builder()
-                .securityReferences(Collections.singletonList(new SecurityReference("OAuth2.0", scoper())))
-                .forPaths(PathSelectors.regex("^(?!auth).*$"))
-                .build();
-    }
-
-    private SecurityScheme securityScheme() {
-        GrantType grantType = new ResourceOwnerPasswordCredentialsGrant(tokenUrl);
-        return new OAuthBuilder()
-                .name("OAuth2.0")
-                .grantTypes(Collections.singletonList(grantType))
-                .scopes(Arrays.asList(scoper()))
-                .build();
-    }
-
-    private AuthorizationScope[] scoper() {
-        return new AuthorizationScope[]{
-                new AuthorizationScope(SCOPE, "webapp scope is trusted!")
-        };
-    }
-}
