--- learn-cloud-zuul/src/main/java/com/billow/zuul/redis/RedisConfig.java
@@ -1,52 +1,0 @@
-package com.billow.zuul.redis;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisPassword;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceClientConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.StringRedisTemplate;
-
-/**
- * redis 配置文件
- *
- * @author LiuYongTao
- * @date 2019/7/29 9:13
- */
-@Configuration
-@EnableAutoConfiguration
-public class RedisConfig {
-
-    private static Logger logger = LoggerFactory.getLogger(RedisConfig.class);
-
-    @Autowired
-    private RedisProperties redisProperties;
-
-    @Bean
-    public LettuceConnectionFactory getConnectionFactory() {
-        RedisStandaloneConfiguration redisStandaloneConfiguration = new RedisStandaloneConfiguration();
-        redisStandaloneConfiguration.setDatabase(redisProperties.getDatabase());
-        redisStandaloneConfiguration.setHostName(redisProperties.getHost());
-        redisStandaloneConfiguration.setPort(redisProperties.getPort());
-        redisStandaloneConfiguration.setPassword(RedisPassword.of(redisProperties.getPassword()));
-
-        LettuceClientConfiguration.LettuceClientConfigurationBuilder lettuceClientConfigurationBuilder = LettuceClientConfiguration.builder();
-        LettuceConnectionFactory factory = new LettuceConnectionFactory(redisStandaloneConfiguration,
-                lettuceClientConfigurationBuilder.build());
-
-        logger.info("JedisConnectionFactory bean init success.");
-        return factory;
-    }
-
-    @Bean
-    public RedisTemplate<?, ?> redisTemplate() {
-        return new StringRedisTemplate(getConnectionFactory());
-    }
-}
