+++ 配置文件备份/cloud-config-core/cloud-config-dev.properties
@@ -10,14 +10,14 @@ config.register.namespace=${spring.cloud.nacos.config.namespace}
 # \u6CE8\u518C\u4E2D\u5FC3\u5730\u5740
 config.register.serverAddr=http://${spring.cloud.nacos.config.server-addr}
 ####################################################################################################
-################################################## learn-cloud-zuul ################################
+################################################## learn-cloud-gateway ################################
 ####################################################################################################
 #\u5E94\u7528\u7AEF\u53E3
-config.zuul.server.port=8771
+config.gateway.server.port=8771
 # \u79FB\u52A8 /usr/app/app
 # pc /usr/app/web
-config.zuul.static.locations=file:///D:/GitHub/learn/learn-shop-ui-admin/dist
-#config.zuul.static.locations=file:///D:/usr/app
+config.gateway.static.locations=file:///D:/GitHub/learn/learn-shop-ui-admin/dist
+#config.gateway.static.locations=file:///D:/usr/app
 ####################################################################################################
 ################################################## learn-cloud-turbine #############################
 ####################################################################################################
