+++ 配置文件备份/cloud-config-core/cloud-config-pro.properties
@@ -10,7 +10,7 @@ config.register.namespace=${spring.cloud.nacos.config.namespace}
 # \u6CE8\u518C\u4E2D\u5FC3\u5730\u5740
 config.register.serverAddr=http://${spring.cloud.nacos.config.server-addr}
 ####################################################################################################
-################################################## learn-cloud-zuul ################################
+################################################## learn-cloud-gateway ################################
 ####################################################################################################
 #\u5E94\u7528\u7AEF\u53E3
 config.zuul.server.port=8771
